Move class and change import path.
